backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
